Development

See merge request eagles-creative/opencreative!57

+++ src/main/java/ua/mcchickenstudio/opencreative/utils/FileUtils.java
@@ -462,6 +462,7 @@ public class FileUtils {
      */
     public static void deleteFolder(File directory) {
         if (!directory.exists()) return;
+        if (directory.equals(Bukkit.getWorldContainer())) return;
         try {
             org.apache.commons.io.FileUtils.deleteDirectory(directory);
         } catch (IOException error) {
@@ -671,8 +672,8 @@ public class FileUtils {
 
     public static String getPlanetIdFromName(String name) {
         return name
-                .replace(Bukkit.getServer().getWorldContainer() + File.separator,"")
-                .replace("planets" + File.separator + "planet","");
+                .replace(Bukkit.getServer().getWorldContainer().getPath().replace("\\","/") + "/","")
+                .replace("planets/planet","");
     }
 
 }
